refactor(#60): split dashboard-files
Split the vue-files of the dashboard by moving
the modals into separate files, to make the code
cleaner and easier to update.
