+++ src/dashboard/app/src/components/workload/task/task_overview.vue
@@ -76,82 +76,12 @@
                 <p v-else>No tasks found</p>
             </div>
         </div>
-
-        <!-- Add Task Modal -->
-        <div
-            v-if="showAddModal"
-            class="modal-overlay"
-            @click.self="cancelAddModal"
-        >
-            <div class="modal task-create-modal">
-                <!-- Modal topbar -->
-                <div class="modal-topbar">
-                    <span>Create task</span>
-                </div>
-                <div class="modal-content">
-                    <input
-                        v-model="newTask.taskName"
-                        type="text"
-                        placeholder="Task-Name"
-                        required
-                    />
-                    <div>
-                        <label>Task template:</label>
-                        <textarea
-                            id="template_input"
-                            v-model="newTask.taskTemplate"
-                            type="text"
-                            required
-                        ></textarea>
-                    </div>
-                </div>
-
-                <div class="modal-bottombar">
-                    <div class="modal-actions">
-                        <button class="icon-button" @click="acceptAddModal">
-                            <img :src="icons.acceptIcon" alt="Accept" />
-                        </button>
-                        <button class="icon-button" @click="cancelAddModal">
-                            <img :src="icons.cancelIcon" alt="Cancel" />
-                        </button>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <!-- Delete Confirmation Modal -->
-        <div
-            v-if="showDeleteModal"
-            class="modal-overlay"
-            @click.self="cancelDeleteModal"
-        >
-            <div class="modal task-delete-modal">
-                <div class="modal-topbar">
-                    <span>Delete task</span>
-                </div>
-                <div class="modal-content">
-                    <p>Are you sure you want to delete?</p>
-                    <strong>Task: {{ taskToDelete?.uuid }}</strong>
-                </div>
-
-                <div class="modal-bottombar">
-                    <div class="modal-actions">
-                        <button class="icon-button" @click="acceptDeleteModal">
-                            <img :src="icons.acceptIcon" alt="Accept" />
-                        </button>
-                        <button class="icon-button" @click="cancelDeleteModal">
-                            <img :src="icons.cancelIcon" alt="Cancel" />
-                        </button>
-                    </div>
-                </div>
-            </div>
-        </div>
     </div>
 </template>
 
 <script setup lang="ts">
 import { ref, onMounted, onBeforeUnmount, inject } from "vue";
-import api from "../../api";
+import api from "../../../api";
 import "primevue/resources/themes/saga-blue/theme.css"; // or any other theme
 import ProgressBar from "primevue/progressbar";
 
@@ -163,10 +93,6 @@ const tasks = ref<{ uuid: string; taskName: string }[]>([]);
 const showAddModal = ref(false);
 const showDeleteModal = ref(false);
 const openDropdown = ref<string | null>(null);
-const newTask = ref({
-    taskTemplate: "",
-    taskName: "",
-});
 const passwordError = ref("");
 const taskToDelete = ref<{ uuid: string; taskName: string } | null>(null);
 const icons = inject<{ acceptIcon: string; cancelIcon: string }>("icons")!;
@@ -213,70 +139,6 @@ function handleClickOutside(event: MouseEvent) {
 }
 
 //=============================================================================
-// Add task modal
-//=============================================================================
-function openAddModal() {
-    showAddModal.value = true;
-}
-function cancelAddModal() {
-    showAddModal.value = false;
-    newTask.value.taskTemplate = "";
-    newTask.value.taskName = "";
-}
-
-async function acceptAddModal() {
-    try {
-        const token = localStorage.getItem("jwtToken");
-        await api.hanami_api.post(
-            "/v1alpha/task",
-            {
-                name: newTask.value.taskName,
-                template: newTask.value.taskTemplate,
-            },
-            {
-                headers: { Authorization: `Bearer ${token}` },
-            },
-        );
-        await fetchTasks();
-        cancelAddModal();
-    } catch (err) {
-        passwordError.value = err;
-        console.error("Failed to create task", err);
-    }
-}
-
-//=============================================================================
-// Delete modal
-//=============================================================================
-function openDeleteModal(task: { uuid: string; taskName: string }) {
-    taskToDelete.value = task;
-    showDeleteModal.value = true;
-    openDropdown.value = null;
-}
-function cancelDeleteModal() {
-    showDeleteModal.value = false;
-    taskToDelete.value = null;
-    openDropdown.value = null; // close any open action dropdown
-}
-async function acceptDeleteModal() {
-    if (!taskToDelete.value) return;
-    try {
-        const token = localStorage.getItem("jwtToken");
-        await api.hanami_api.delete(
-            `/v1alpha/task/${taskToDelete.value.uuid}`,
-            {
-                headers: { Authorization: `Bearer ${token}` },
-            },
-        );
-        await fetchTasks();
-    } catch (err) {
-        console.error("Failed to delete task", err);
-    } finally {
-        cancelDeleteModal();
-    }
-}
-
-//=============================================================================
 // Listener
 //=============================================================================
 onMounted(fetchTasks);
@@ -291,16 +153,6 @@ onBeforeUnmount(() => {
 </script>
 
 <style scoped>
-.task-create-modal {
-    height: 35rem;
-    width: 30rem;
-}
-
-.task-delete-modal {
-    height: 16rem;
-    width: 20rem;
-}
-
 .overview-table td:nth-child(3) {
     width: 10rem;
 }
